requirements.txt
google-adk==1.5.0
lunar_python==1.4.4
google-cloud-aiplatform==1.100.0
